feat(explorer): refactor account list to component
