+++ apps/explorer/src/app/routes/parties/id/Party-assets.graphql
@@ -1,3 +1,28 @@
+fragment ExplorerPartyAssetsAccounts on AccountBalance {
+  asset {
+    name
+    id
+    decimals
+    symbol
+    source {
+      __typename
+      ... on ERC20 {
+        contractAddress
+      }
+    }
+  }
+  type
+  balance
+  market {
+    id
+    tradableInstrument {
+      instrument {
+        name
+      }
+    }
+  } 
+}
+
 query ExplorerPartyAssets($partyId: ID!) {
   partiesConnection(id: $partyId) {
     edges {
@@ -21,28 +46,7 @@ query ExplorerPartyAssets($partyId: ID!) {
         accountsConnection {
           edges {
             node {
-              asset {
-                name
-                id
-                decimals
-                symbol
-                source {
-                  __typename
-                  ... on ERC20 {
-                    contractAddress
-                  }
-                }
-              }
-              type
-              balance
-              market {
-                id
-                tradableInstrument {
-                  instrument {
-                    name
-                  }
-                }
-              }
+              ... ExplorerPartyAssetsAccounts 
             }
           }
         }
